cdk/requirements.txt
aws-cdk.core
aws-cdk.aws-lambda
aws-cdk.aws-iam
aws-cdk.aws-events
aws-cdk.aws-events-targets

